+++ halotools/mock_observables/__init__.py
@@ -12,7 +12,6 @@ from .marked_tpcf import *
 from .rp_pi_tpcf import *
 from .wp import *
 from .s_mu_tpcf import *
-from .delta_sigma import *
 from .group_identification import *
 from .mock_survey import *
 from .tpcf_multipole import *
