--- halotools/mock_observables/test_clustering/__init__.py
@@ -1,12 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
-
-from .test_tpcf import *
-from .test_tpcf_jackknife import *
-from .test_wp import *
-from .test_s_mu_tpcf import *
-from .test_marked_tpcf import *
-from .test_rp_pi_tpcf import *
-from .test_delta_sigma import *
-from .test_tpcf_one_two_halo import *
