+++ halotools/mock_observables/two_point_clustering/delta_sigma.py
@@ -7,10 +7,10 @@ from scipy.interpolate import InterpolatedUnivariateSpline
 from scipy import integrate
 from warnings import warn
 
-from .tpcf import tpcf
-from .mock_observables_helpers import (get_num_threads, get_separation_bins_array, 
+from ..tpcf import tpcf
+from ..mock_observables_helpers import (get_num_threads, get_separation_bins_array, 
     get_period, enforce_sample_respects_pbcs, enforce_sample_has_correct_shape)
-from .clustering_helpers import verify_tpcf_estimator
+from ..clustering_helpers import verify_tpcf_estimator
 
 __all__ = ['delta_sigma']
 __author__ = ['Duncan Campbell']
